[ORGA] [SR] Move recommendation analysis tests in correct folder
